+++ relay-dynamic-config/src/project.rs
@@ -1,5 +1,4 @@
 use std::collections::BTreeSet;
-use std::sync::Arc;
 
 use relay_auth::PublicKey;
 use relay_filter::FiltersConfig;
@@ -16,7 +15,7 @@ use serde_json::Value;
 use crate::metrics::{
     MetricExtractionConfig, SessionMetricsConfig, TaggingRule, TransactionMetricsConfig,
 };
-use crate::{ErrorBoundary, FeatureSet, GlobalConfig};
+use crate::{ErrorBoundary, FeatureSet};
 
 /// Dynamic, per-DSN configuration passed down from Sentry.
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -105,32 +104,6 @@ impl Default for ProjectConfig {
             tx_name_ready: false,
             span_description_rules: None,
         }
-    }
-}
-
-impl ProjectConfig {
-    /// Measurements are usually defined globally, but can be overriden per DNS. We therefore
-    /// first check if the project config has a value, if not we fall back on global.
-    pub fn _measurements<'a>(
-        &'a self,
-        global_config: &'a Arc<GlobalConfig>,
-    ) -> Option<&MeasurementsConfig> {
-        self.measurements
-            .as_ref()
-            .or(global_config.measurements.as_ref())
-    }
-
-    /// There's three types of tagging rules, one which is defined per dsn, one that is defined
-    /// globally for every dsn, and one that is only sometimes defined by the dsn. this method
-    /// will merge them together in the right order.
-    /// TODO(tor): Implement the maybe-override values logic.
-    pub fn _metric_conditional_tagging<'a>(
-        &'a self,
-        global_config: &'a Arc<GlobalConfig>,
-    ) -> impl Iterator<Item = &TaggingRule> {
-        self.metric_conditional_tagging
-            .iter()
-            .chain(global_config.metric_conditional_tagging.iter())
     }
 }