+++ relay-server/src/statsd.rs
@@ -536,10 +536,8 @@ pub enum RelayCounters {
     ///  - `sdk`: The name of the Sentry SDK sending the transaction. This tag is only set for
     ///    Sentry's SDKs and defaults to "proprietary".
     OpenTelemetryEvent,
-    /// Successfully read from the global config.
-    GlobalConfigReadSuccess,
-    /// Successfully read from the global config.
-    GlobalConfigReadFailed,
+    /// Reading the global config from the RwLock
+    GlobalConfigRead,
 }
 
 impl CounterMetric for RelayCounters {
@@ -574,9 +572,7 @@ impl CounterMetric for RelayCounters {
             RelayCounters::MetricBucketsParsingFailed => "metrics.buckets.parsing_failed",
             RelayCounters::MetricsTransactionNameExtracted => "metrics.transaction_name",
             RelayCounters::OpenTelemetryEvent => "event.opentelemetry",
-            // Just temporary to check if infra for global config works before we implement it
-            RelayCounters::GlobalConfigReadSuccess => "global_config.read_success",
-            RelayCounters::GlobalConfigReadFailed => "global_config.read_failed",
+            RelayCounters::GlobalConfigRead => "global_config.read",
         }
     }
 }