+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 **Internal**:
 
+- Add new service for fetching global configs. ([#2320](https://github.com/getsentry/relay/pull/2320))
 - Add support for `sampled` field in the DSC and error tagging. ([#2290](https://github.com/getsentry/relay/pull/2290))
 - Move span tag extraction from metrics to normalization. ([#2304](https://github.com/getsentry/relay/pull/2304))